requirements.txt
flask
flask-sock
groq
supabase
sentence-transformers
python-dotenv
ddgs
requests
pygame
edge-tts